script/requirements.txt
playwright
pyyaml
aiohttp